Improve extensibility of Zxcvbn and Matching
Previously the only way how to provide custom dictionaries was by using
`sanitizedInputs` argument of `Zxcvbn#measure(String, List<String>)` method.
The disadvantage was that the custom dictionary was reprocessed for every call
and for large custom dictionary it could incur considerable overhead.

To improve the situation, Zxcvbn class is modified to allow supplying custom
instance of Matching class and also Matching class is changed to allow the
needed level of extensibility.

This extensibility changes maintain full backward compatibility.